backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
sqlmodel==0.0.22
aiosqlite==0.20.0
google-generativeai==0.8.5
playwright==1.51.0
httpx==0.28.1
beautifulsoup4==4.13.3
lxml==5.3.2
python-multipart==0.0.20
pydantic-settings==2.8.1
